ptt_scraper/requirements.txt
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==36.0.2
cssselect==1.2.0
filelock==3.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
jmespath==1.0.1
lxml==5.2.1
mysql-connector-python==8.4.0
packaging==24.0
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==22.0.0
queuelib==1.7.0
requests==2.31.0
requests-file==2.0.0
Scrapy==2.6.1
service-identity==24.1.0
six==1.16.0
tldextract==5.1.2
Twisted==22.10.0
typing-extensions==4.11.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.3
